authentication responsibility moved to separate service. API gateway is now a simple gateway.
