+++ gate.go
@@ -3,11 +3,10 @@ package main
 import (
 	"encoding/json"
 	"errors"
-	"fmt"
 	"net/http"
 	"os"
 	"regexp"
-	"text/template"
+	"strings"
 
 	"golang.org/x/net/http2"
 
@@ -15,19 +14,21 @@ import (
 	"github.com/nats-io/nats"
 	"github.com/sebest/logrusly"
 	"github.com/sohlich/natsproxy"
-	"github.com/suricatatalk/gate/auth"
+	"github.com/suricatatalk/gate/jwt"
+	"github.com/suricatatalk/gate/storage"
+	"github.com/suricatatalk/guardian/auth"
 	"github.com/suricatatalk/mail/client"
 
 	"github.com/kelseyhightower/envconfig"
 
 	// "github.com/markbates/goth"
-	"github.com/markbates/goth/gothic"
+
 	// "github.com/markbates/goth/providers/twitter"
 )
 
 const (
 	ServiceName = "gateway"
-	TokenHeader = "X-AUTH"
+	TokenHeader = "X-Auth"
 
 	//Configuration keys
 	KeyLogly = "LOGLY_TOKEN"
@@ -36,11 +37,10 @@ const (
 var (
 	ErrNoServiceInUrl          = errors.New("No service definition in url")
 	ErrInavelidActivationToken = errors.New("Invalid activation token")
-	authProvider               auth.AuthProvider
-	passManager                auth.PasswordManager
 	actiavteUserRegex          = regexp.MustCompile(".*\\/activate\\/")
 	passwordResetRegex         = regexp.MustCompile(".*\\/resetpassword\\/")
 	appCfg                     *AppConfig
+	tokenStorage               storage.DataStorage
 )
 
 var (
@@ -56,9 +56,10 @@ type AppConfig struct {
 	Domain string `default:"suricata.cleverapps.io"`
 }
 
-type MgoConfig struct {
-	URI string `default:"127.0.0.1:27017"`
-	DB  string `default:"surikata"`
+type RedisConfig struct {
+	URI      string `default:"127.0.0.1:6379"`
+	Password string `default:""`
+	DB       int64  `default:"0"`
 }
 
 type NatsConfig struct {
@@ -66,12 +67,12 @@ type NatsConfig struct {
 }
 
 // loadConfiguration loads the configuration of application
-func loadConfiguration(app *AppConfig, mgo *MgoConfig, nats *NatsConfig) {
+func loadConfiguration(app *AppConfig, rdis *RedisConfig, nats *NatsConfig) {
 	err := envconfig.Process(ServiceName, app)
 	if err != nil {
 		log.Panicln(err)
 	}
-	err = envconfig.Process("mongodb", mgo)
+	err = envconfig.Process("redis", rdis)
 	if err != nil {
 		log.Panicln(err)
 	}
@@ -94,37 +95,29 @@ func main() {
 	configureSocial()
 	// Load all configuration
 	appCfg = &AppConfig{}
-	mgoCfg := &MgoConfig{}
+	rdisCfg := &RedisConfig{}
 	natsCfg := &NatsConfig{}
-	loadConfiguration(appCfg, mgoCfg, natsCfg)
+	loadConfiguration(appCfg, rdisCfg, natsCfg)
 
 	// Service discovery config
-	log.Infoln("Loading configuration for ETCD client")
 
-	initMail()
-	var mailErr error
-	mailClient, mailErr = client.NewNatsMailClient(natsCfg.Endpoint)
-	if mailErr != nil {
-		log.Errorf("Cannot initialize mail client: %s", mailErr.Error())
-	}
 	//Mongo configuration
-	log.Infoln("Loading configuration of MongoDB")
-	mgoStorage := auth.NewMgoStorage()
-	mgoStorage.ConnectionString = mgoCfg.URI
-	mgoStorage.Database = mgoCfg.DB
-	err := mgoStorage.OpenSession()
+	log.Infoln("Loading configuration of Redis")
+	rdisStorage := storage.NewRedisStorage()
+	rdisStorage.URL = rdisCfg.URI
+	rdisStorage.Password = rdisCfg.Password
+	rdisStorage.Database = rdisCfg.DB
+	tokenStorage = rdisStorage
+	err := tokenStorage.OpenSession()
+	defer tokenStorage.CloseSession()
 	if err != nil {
 		log.Panic(err)
 	}
-	log.Infoln("Initializing auth provider")
-	mgoAuthProvider := auth.NewAuthProvider(mgoStorage)
-	authProvider = mgoAuthProvider
-	passManager = mgoAuthProvider
 
-	log.Infoln("Initializing reverse proxy")
-
+	log.Infoln("Initializing NATS proxy")
 	proxyConn, _ := nats.Connect(nats.DefaultURL)
 	multiProxy, err := natsproxy.NewNatsProxy(proxyConn)
+	multiProxy.AddHook(loginHook)
 	defer proxyConn.Close()
 	if err != nil {
 		log.Panic("Cannot initialize NATS proxy")
@@ -133,21 +126,15 @@ func main() {
 	log.Infoln("Registering handlers")
 	//Handle login and register
 	mux := http.NewServeMux()
-	mux.HandleFunc("/activate/", activateHandler)
-	mux.HandleFunc("/login", loginHandler)
-	mux.HandleFunc("/register", registerHandler)
-
-	mux.HandleFunc("/requestpasswordreset", requestPasswordResetHandler)
-	mux.HandleFunc("/resetpassword/", passwordResetHandler)
-	// mux.Get("/auth/{provider}/callback", handleSocialLogin)
-	// mux.Get("/auth/{provider}", gothic.BeginAuthHandler)
-	//else handle via proxy
 	log.Infoln("Start listening on " + appCfg.Port)
-
 	mux.HandleFunc("/", func(rw http.ResponseWriter, req *http.Request) {
+		log.Println("Handling")
 		transformToken(req)
+		log.Println(req.Header)
 		multiProxy.ServeHTTP(rw, req)
 	})
+
+	//Init HTTP2 server
 	srv := &http.Server{
 		Addr:    ":" + appCfg.Port,
 		Handler: mux,
@@ -159,20 +146,23 @@ func main() {
 	}
 }
 
-func initMail() {
-	subjectTemp, _ := template.New("activate_subject").Parse("Suricata: Registration confirmation")
-	messageTemp, _ := template.New("activate_message").Parse("Please confirm the registration on Suricata Talk website with click on this link {{.ConfirmationLink}}")
-	activationComposer = &client.SuricataMessageComposer{
-		subjectTemp,
-		messageTemp,
+func loginHook(urlPath string, resp *natsproxy.Response) {
+	if !strings.Contains(urlPath, "/login") {
+		return
 	}
-
-	subjectTemp, _ = template.New("passreset_subject").Parse("Suricata: Password reset")
-	messageTemp, _ = template.New("passreset_message").Parse("Reset the password on following link {{.ResetLink}}")
-	passResetComposer = &client.SuricataMessageComposer{
-		subjectTemp,
-		messageTemp,
+	userJSON := resp.Header.Get(TokenHeader)
+	user := auth.User{}
+	json.Unmarshal([]byte(userJSON), &user)
+	jwtToken, err := jwt.GenerateJwtToken(user)
+	if err != nil {
+		log.Error(err)
 	}
+	token := storage.NewToken(user, jwtToken)
+	err = tokenStorage.InsertToken(token)
+	if err != nil {
+		log.Error(err)
+	}
+	resp.Header.Set(TokenHeader, token.RefToken)
 }
 
 func configureSocial() {
@@ -185,137 +175,28 @@ func transformToken(req *http.Request) {
 		return
 	}
 	log.Infof("Getting security token %s", refToken)
-	valToken, err := authProvider.ValueToken(refToken)
+	valToken, err := tokenStorage.TokenByRefToken(refToken)
 
-	if err == nil && len(valToken) != 0 {
+	if err == nil && len(valToken.JwtToken) != 0 {
 		log.Infof("Setting valueToken %s", valToken)
-		req.Header.Set(TokenHeader, valToken)
-	}
-
-}
-
-func loginHandler(rw http.ResponseWriter, req *http.Request) {
-	decoder := json.NewDecoder(req.Body)
-	payload := struct {
-		Email    string
-		Password string
-	}{}
-
-	decodeErr := decoder.Decode(&payload)
-	if decodeErr != nil {
-		http.Error(rw, decodeErr.Error(), http.StatusBadRequest)
-		return
-	}
-
-	token, err := authProvider.SignIn(payload.Email, payload.Password)
-	if err != nil && err != auth.ErrPasswordNotMatch {
-		http.Error(rw, err.Error(), http.StatusInternalServerError)
-		return
-	} else if err == auth.ErrPasswordNotMatch {
-		http.Error(rw, err.Error(), http.StatusForbidden)
-		return
-	}
-
-	rw.Header().Set(TokenHeader, token)
-}
-
-func registerHandler(rw http.ResponseWriter, req *http.Request) {
-	decoder := json.NewDecoder(req.Body)
-	user := auth.NewInactiveUser()
-	decodeErr := decoder.Decode(&user)
-	if decodeErr != nil {
-		http.Error(rw, decodeErr.Error(), http.StatusBadRequest)
-		return
-	}
-	signUpErr := authProvider.SignUp(user)
-	if signUpErr != nil {
-		http.Error(rw, signUpErr.Error(), http.StatusBadRequest)
-		return
-	}
-
-	actToken, err := authProvider.RequestUserActivationFor(user.Email)
-	if err != nil {
-		http.Error(rw, "Internal server error", http.StatusInternalServerError)
-	}
-	sendActivationMailToUser(user.Email, actToken)
-
-	jsonVal, _ := json.Marshal(user)
-	rw.Write(jsonVal)
-}
-
-func activateHandler(rw http.ResponseWriter, req *http.Request) {
-	log.Println("Activate handler")
-	token := actiavteUserRegex.ReplaceAllString(req.URL.Path, "")
-	if len(token) != 36 {
-		http.Error(rw, ErrInavelidActivationToken.Error(), http.StatusBadRequest)
-	}
-	authProvider.ActivateUser(token)
-}
-
-func requestPasswordResetHandler(rw http.ResponseWriter, req *http.Request) {
-	log.Println("Request password reset")
-	email := req.FormValue("email")
-	if len(email) == 0 {
-		log.Infoln("Parameter \"email\" not found")
-		http.Error(rw, "Parameter not found", http.StatusBadRequest)
-		return
-	}
-	token, err := passManager.RequestPasswordResetFor(email)
-	if err != nil {
-		log.Error(err)
-		http.Error(rw, "Internal server error", http.StatusInternalServerError)
-		return
-	}
-	//TODO send mail with token
-	err = sendPasswordResetMail(email, token)
-	if err != nil {
-		log.Error("Could not send password reset for mail: %s", email)
-		http.Error(rw, "Internal server error", http.StatusInternalServerError)
-		return
+		req.Header.Set(TokenHeader, valToken.JwtToken)
 	}
 }
 
-func passwordResetHandler(rw http.ResponseWriter, req *http.Request) {
-	tkn := passwordResetRegex.ReplaceAllString(req.URL.Path, "")
-	pass := req.FormValue("password")
-	err := passManager.ResetPasswordBy(tkn, pass)
-	if err == auth.ErrResetTokenExpired {
-		http.Error(rw, "Expired", http.StatusForbidden)
+// func handleSocialLogin(rw http.ResponseWriter, req *http.Request) {
+// 	log.Println(gothic.GetState(req))
+// 	socialUser, err := gothic.CompleteUserAuth(rw, req)
+// 	if err != nil {
+// 		log.Println(err)
+// 		http.Error(rw, err.Error(), http.StatusBadRequest)
+// 		return
+// 	}
 
-	} else if err != nil {
-		log.Error(err)
-		http.Error(rw, "Internal server error", http.StatusInternalServerError)
-	}
-}
+// 	user := auth.User{}
+// 	user.UserID = socialUser.UserID
+// 	user.Email = socialUser.Email
 
-func handleSocialLogin(rw http.ResponseWriter, req *http.Request) {
-	log.Println(gothic.GetState(req))
-	socialUser, err := gothic.CompleteUserAuth(rw, req)
-	if err != nil {
-		log.Println(err)
-		http.Error(rw, err.Error(), http.StatusBadRequest)
-		return
-	}
-
-	user := auth.User{}
-	user.UserID = socialUser.UserID
-	user.Email = socialUser.Email
-
-	log.Println(socialUser.UserID)
-	log.Println(socialUser.AccessToken)
-	log.Println(socialUser.NickName)
-}
-
-func sendActivationMailToUser(email, token string) error {
-	messageStruct := struct{ ConfirmationLink string }{fmt.Sprintf("http://%s/activate/%s", appCfg.Domain, token)}
-	subject := activationComposer.ComposeSubject(struct{}{})
-	message := activationComposer.ComposeMessage(messageStruct)
-	return mailClient.SendMail(email, subject, message)
-}
-
-func sendPasswordResetMail(email, token string) error {
-	messageStruct := struct{ ResetLink string }{fmt.Sprintf("http://%s/resetpassword/%s", appCfg.Domain, token)}
-	subject := passResetComposer.ComposeSubject(struct{}{})
-	message := passResetComposer.ComposeMessage(messageStruct)
-	return mailClient.SendMail(email, subject, message)
-}
+// 	log.Println(socialUser.UserID)
+// 	log.Println(socialUser.AccessToken)
+// 	log.Println(socialUser.NickName)
+// }
